backend/requirements.txt
fastapi>=0.65.0
uvicorn>=0.14.0
python-multipart>=0.0.5
pillow>=8.3.0
numpy>=1.21.0
opencv-python>=4.5.0
pyzbar>=0.1.8
pytesseract>=0.3.8
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.5.0
python-jose==3.3.0
python-dotenv==1.0.0
requests>=2.26.0
scikit-image>=0.18.0
scikit-learn>=0.24.0 